Feature/ completed, only lefts a correct SQL query to retrive the correct data

+++ src/main/java/com/rulesAPI/repositorios/ConclusionRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.CrudRepository;
  *
  * @author laloses
  */
-public interface ConclusionRepository extends CrudRepository<Conclusion, Integer>{
+public interface ConclusionRepository extends CrudRepository<Conclusion, Long>{
     
 }
