+++ src/main/java/com/rulesAPI/repositorios/HechoRepository.java
@@ -4,9 +4,8 @@ import com.rulesAPI.modelos.entidades.Hecho;
 import org.springframework.data.repository.CrudRepository;
 
 /**
- *
  * @author laloses
  */
-public interface HechoRepository extends CrudRepository<Hecho, Integer>{
+public interface HechoRepository extends CrudRepository<Hecho, Long>{
     
 }
